feat(trans): Delay render for snappier user xp
